Catch up with -current.

+++ sys/dev/pci/if_ntwoc_pci.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: if_ntwoc_pci.c,v 1.4 2000/12/28 22:59:13 sommerfeld Exp $	*/
+/*	$NetBSD: if_ntwoc_pci.c,v 1.4.2.1 2001/08/24 00:10:05 nathanw Exp $	*/
 
 /*
  * Copyright (c) 1998 Vixie Enterprises
@@ -718,7 +718,7 @@ ntwoc_pci_setup_dma(struct sca_softc *sc)
 	 */
 	if (sc->scu_allocsize != addroff)
 		printf("ERROR:  scu_allocsize != addroff: %lu != %lu\n",
-		       sc->scu_allocsize, addroff);
+		       (u_long)sc->scu_allocsize, addroff);
 }
 
 #if __NetBSD_Version__ >= 104160000
